detect MTS

+++ chop.py
@@ -32,13 +32,17 @@ def filterTimes(l, threshold):
  return times
 
  
-def getWavPath(videoPath):
- return re.match('(.*[/\][\w_-]+\.)\w+', videoPath).group(1) + "wav"
+def changeExtension(videoPath, ext):
+ return re.match('(.*[/\][\w_-]+\.)\w+', videoPath).group(1) + ext
 
 
 def getChoppedPath(videoPath, i):
  m = re.match('(.*[/\][\w_-]+)(\.\w+)', videoPath)
  return m.group(1) + "_" + str(i).zfill(3) + m.group(2)
+
+
+def isMTS(videoPath):
+ return re.match("(?i).*\.mts$", videoPath) is not None
 
 
 def timeify(time):
@@ -56,14 +60,19 @@ def chopVideo(videoPath, times):
   sp.call(["avconv", "-i", videoPath, "-y", "-ss", timeify(start), "-t", timeify(duration), "-codec", "copy", choppedPath])
  sp.call(["avconv", "-i", videoPath, "-y", "-ss", timeify(times.pop()), "-codec", "copy", getChoppedPath(videoPath, len(times) + 1)])
 
-def chopify(videoPath, freq1, freq2, freq3): 
- wavPath = getWavPath(videoPath)
- sp.call(["avconv", "-i", videoPath, "-ab", "160k", "-ac", "1", "-ar", "160000", "-vn", wavPath])
- (fs,data) = read(wavPath)
- os.remove(wavPath)
+def chopify(videoPath, freq1, freq2, freq3):
+ if (isMTS(videoPath)):
+  print "IT'S MTS!"
+ else:
+  wavPath = changeExtension(videoPath, "wav")
+  sp.call(["avconv", "-i", videoPath, "-ab", "160k", "-ac", "1", "-ar", "160000", "-vn", wavPath])
+  (fs,data) = read(wavPath)
+  os.remove(wavPath)
 
- l = rateSingle(data, fs, freq1, freq2, freq3)
- thresh = sorted(l, key = lambda x: x[1], reverse=True)[0][1]/2
- times = filterTimes(l, thresh)
- chopVideo(videoPath, times)
+  l = rateSingle(data, fs, freq1, freq2, freq3)
+  thresh = sorted(l, key = lambda x: x[1], reverse=True)[0][1]/2
+  times = filterTimes(l, thresh)
+  chopVideo(videoPath, times)
+
+
